requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
SQLAlchemy==2.0.34
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
psutil==6.0.0
GitPython==3.1.43
waitress==3.0.0
